Add uv to Dependabot config (#10)
We haven’t been having Dependabot keep Python things up-to-date, partially because Dependabot had incomplete uv support when this was first written, and partially because this project had been pretty inactive. People are digging in on the problems this is solving again now, so we should really make sure we are keeping better up-to-date.