Iterate over all DB query results (#17)
For some reason, we weren't previously handling paginated responses from Notion when querying for all the existing DB entries, leading to lots of duplicated pages/rules being added. I think I must have left this as a to-do early on (but never marked it as such) since I expected this would get feedback and incur lots of changes quickly. That never happened (people are only just starting to seriously use this tool now), and at some point we accumulated more Notion DB entries than fit in a single response, at which point we started creating dupes because the script would have an incomplete list of rules it had already recorded. :(

Anyway, this fixes the issue!